requirements.txt
Django>=3.2.4,<4
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt<=5.2.0
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
django-extensions<3.2.1
django-cors-headers<=3.13